requirements.txt
asgiref==3.5.2
boto3==1.26.31
botocore==1.29.31
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
Django==4.1.4
django-storages==1.13.2
djangorestframework==3.14.0
idna==3.4
jmespath==1.0.1
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.13
